--- a/compile/DLL/compile64.sh
+++ b/compile/DLL/compile64.sh
@@ -1,4 +1,12 @@
 #!/bin/sh
 
-CXX=i686-w64-mingw32-g++ make clean
-CXX=i686-w64-mingw32-g++ make
+COPTS=''
+
+for arg in "$@"; do
+    if [ $arg = "--optimize" ]; then
+        COPTS='-msse2 -ffast-math -DMOORDYN_SINGLEPRECISSION'
+    fi
+done
+
+CXX=i686-w64-mingw32-g++ CC=i686-w64-mingw32-gcc make clean
+CXX=i686-w64-mingw32-g++ CC=i686-w64-mingw32-gcc make COPTS="$COPTS" -j