--- a/source/Eigen/Dense
+++ b/source/Eigen/Dense
@@ -1,0 +1,7 @@
+#include "Core"
+#include "LU"
+#include "Cholesky"
+#include "QR"
+#include "SVD"
+#include "Geometry"
+#include "Eigenvalues"