--- a/source/Eigen/Eigen
+++ b/source/Eigen/Eigen
@@ -1,0 +1,2 @@
+#include "Dense"
+#include "Sparse"