--- a/source/Eigen/src/AccelerateSupport/InternalHeaderCheck.h
+++ b/source/Eigen/src/AccelerateSupport/InternalHeaderCheck.h
@@ -1,0 +1,3 @@
+#ifndef EIGEN_ACCELERATESUPPORT_MODULE_H
+#error "Please include Eigen/AccelerateSupport instead of including headers inside the src directory directly."
+#endif