--- a/source/Eigen/src/Cholesky/InternalHeaderCheck.h
+++ b/source/Eigen/src/Cholesky/InternalHeaderCheck.h
@@ -1,0 +1,3 @@
+#ifndef EIGEN_CHOLESKY_MODULE_H
+#error "Please include Eigen/Cholesky instead of including headers inside the src directory directly."
+#endif