… source/Eigen/Geometry
… source/Eigen/Householder
… source/Eigen/IterativeLinearSolvers
… source/Eigen/Jacobi
… source/Eigen/KLUSupport
… source/Eigen/LU
… source/Eigen/MetisSupport
… source/Eigen/OrderingMethods
… source/Eigen/PaStiXSupport
… source/Eigen/PardisoSupport
… source/Eigen/QR
… source/Eigen/QtAlignedMalloc
… source/Eigen/SPQRSupport
… source/Eigen/SVD
… source/Eigen/Sparse
… source/Eigen/SparseCholesky
… source/Eigen/SparseCore
… source/Eigen/SparseLU
… source/Eigen/SparseQR
… source/Eigen/StdDeque
… source/Eigen/StdList
… source/Eigen/StdVector
… source/Eigen/SuperLUSupport
… source/Eigen/UmfPackSupport
… source/Eigen/src/AccelerateSupport/AccelerateSupport.h
… source/Eigen/src/Cholesky/LDLT.h
… source/Eigen/src/Cholesky/LLT.h
… source/Eigen/src/Cholesky/LLT_LAPACKE.h
… source/Eigen/src/CholmodSupport/CholmodSupport.h
… source/Eigen/src/CholmodSupport/InternalHeaderCheck.h