This PR goes a bit further than https://github.com/mattEhall/MoorDyn/pull/22 , conveniently introducing [Eigen](https://eigen.tuxfamily.org/index.php?title=Main_Page) (the standard for vector & matrix related computations). Thus the code can be terribly simplified, transforming this:

```
for (int J=0; J<3; J++)
	vi[J] = U[i][J] - rd[i][J];
```
			
on this:

`const vec vi = U[i] - rd[i];`

On top of that, Eigen introduces many linear system solvers, which it is also simplifying the code and improving its performance and reliability.

And speaking of performance. The **test_bodies_and_rods** test took

```
real    0m2,133s
user    0m2,120s
sys     0m0,003s
```
to run. With the most aggressive options (which would slightly reduce the accuracy/precision, but are anyway passing the tests flawless) the execution time has became
```
real    0m0,898s
user    0m0,896s
sys     0m0,000s
```

**X2.38 times faster!**

-----

In this PR just the Line object is optimized (and probably I will not upgrade the other parts of the code until I extend https://github.com/mattEhall/MoorDyn/pull/19 to all the entities and get merged).

I should also upgrade the other compilers (All of them should still be working, but it would be nice to have the possibility to set optimization options as well)