@mattEhall 2022-07-06
Finally following up on this @sanguinariojoe:

Really impressive speed boost! 

I guess I have to give in and rely on Eigen now, eh? It does seem very powerful...

My big concern is code readability (especially for less advanced coders like me). I do want MoorDyn's code to be understandable by as many people as possible. In your example above about how the code can be "terribly simplified," I think the original two lines is actually more readable than the simplified one line. The two lines makes it clear that the operation is a vector subtraction and it does not rely on a separate library that the reader would have to look into separately.

I'm not too worried about the above example, but a change I think is a larger impact to code readability is the new vec2array types of functions. I would prefer to avoid those, to keep the code simple to read. In other words, keeping something like
```
for (int I=0; I<3; I++)
	Xd[3*N-3 + 3*i-3 + I] = rd[i][I];
```
instead of
```
moordyn::vec2array(acc,   &Xd[            3 * i - 3]);
```
What do you think? You didn't answer my question in #22 about how much unrolling the for loops helps with computations speed, but I'm guessing it is pretty small (don't compilers often do this anyways?). So we could avoid the vec2array style functions and the speed would still be great, right?

My hope is that we can find a compromise between your many speed improvements and some more simple approaches that will keep the code easy to understand. Maybe something like this:
* We keep the awesome updates to use Eigen
* Can we keep the old for loops rather than the new vec2array style functions?
* With the namespace you added, is it possible to avoid having to write moordyn:: so much?
* I propose we skip the single precision option (and skip the 'real' typedef you added). Numerical problems often occur when using single precision, so it seems like more trouble than it's worth. Unless you know an important use case for single precision.

I noticed you have a new v2 branch that contains a lot of your new features combined. Because so much code has been changed (and sometimes unchanged), I've finally given up keeping track of the commit history. I don't see how it's possible for me to be selective. Instead I'm just looking at the end result and comparing it with older versions. So if it would be easier for us to forget these individual pull requests and just look directly at merging in your final combined version, that is fine with me.