@sanguinariojoe 2022-07-06
Hey Matt! I will be KO for a while because i had a premature baby.

Anyway, some fast comments:

 - As you said, I am preparing a v2 branch where i am packing all the improvements in a quite large pull request.
 - I can provide both kinds of APIs with a bunch of inlines, so that will not be a problem. The new one will be probably more flexible though
 - It will be definitely based on Eigen. After porting all the code, those calls to vec2array and array2vec will be rather scarce. In fact, they will be mainly limited to the C <-> C++ interface. The main benefit of having those functions vs slicing and assigning, vout[i] = vin[i], is that you are letting know other people that a fundamental change/typing/casting is happening.
 - The moordyn:: can be almost always removed, since almost everything is in the moordyn namespace. I like it to be explicit though (it is indeed one of the Python zen statements)
 - The compilers are usually trying to unroll loops when they consider that appropriate. Unfortunately they usually do that really bad... You can look for some info around. On top of that, using loops you are breaking the SIMD operations, i.e. it is very bad for performance. We can further discuss this when i am submitting the large PR, and find a correct balance